fix(api): move ego module under cerebrum (PRD-097, #2516)
ego is a chat surface on top of cerebrum's retrieval engine — it
imports CitationParser, ContextAssemblyService, HybridSearchService,
getEngramService, and several retrieval types from cerebrum. The
post-cerebrum spike framed ego as a sibling, but the runtime
coupling is real and structural.

Move `apps/pops-api/src/modules/ego/` to
`apps/pops-api/src/modules/cerebrum/ego/`. Internal imports adjust:
  - ../cerebrum/<x> → ../<x>     (sibling now)
  - ../core/<x>     → ../../core/<x>
  - ../../<top>     → ../../../<top>

The tRPC public namespace (`ego.*`) is unchanged — apps/pops-api/src/router.ts still mounts `egoRouter` at the
top-level `ego` key, just imported from the new location.
Frontend client code is unaffected.

Resolves all 12 ego→cerebrum entries in
.dependency-cruiser-known-violations.json. Combined with the
two outstanding PRs (#2525, #2526) the baseline drops from 14
to 0.

Closes #2516.

3991/3991 backend tests pass.
