+++ apps/pops-api/src/modules/cerebrum/ego/__tests__/engine.test.ts
@@ -1,6 +1,6 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
-import type { RetrievalResult } from '../../cerebrum/retrieval/types.js';
+import type { RetrievalResult } from '../../retrieval/types.js';
 import type { ChatParams, Message } from '../types.js';
 
 // --- Mocks ---
@@ -15,17 +15,17 @@ const mockHybrid =
     ) => Promise<RetrievalResult[]>
   >();
 
-vi.mock('../../../db.js', () => ({
+vi.mock('../../../../db.js', () => ({
   getDrizzle: () => ({}),
 }));
 
-vi.mock('../../cerebrum/retrieval/hybrid-search.js', () => ({
+vi.mock('../../retrieval/hybrid-search.js', () => ({
   HybridSearchService: class {
     hybrid = mockHybrid;
   },
 }));
 
-vi.mock('../../cerebrum/instance.js', () => ({
+vi.mock('../../instance.js', () => ({
   getEngramService: () => ({
     read: () => {
       throw new Error('No engram in test context');
@@ -44,22 +44,22 @@ vi.mock('@anthropic-ai/sdk', () => ({
   },
 }));
 
-vi.mock('../../../env.js', () => ({
+vi.mock('../../../../env.js', () => ({
   getEnv: (name: string) => {
     if (name === 'ANTHROPIC_API_KEY') return 'test-key';
     return undefined;
   },
 }));
 
-vi.mock('../../../lib/inference-middleware.js', () => ({
+vi.mock('../../../../lib/inference-middleware.js', () => ({
   trackInference: (_params: unknown, fn: () => Promise<unknown>) => fn(),
 }));
 
-vi.mock('../../../lib/ai-retry.js', () => ({
+vi.mock('../../../../lib/ai-retry.js', () => ({
   withRateLimitRetry: (fn: () => Promise<unknown>) => fn(),
 }));
 
-vi.mock('../../../lib/logger.js', () => ({
+vi.mock('../../../../lib/logger.js', () => ({
   logger: {
     info: vi.fn(),
     warn: vi.fn(),
