+++ apps/pops-api/src/modules/cerebrum/ego/__tests__/scope-negotiator.test.ts
