+++ apps/pops-api/src/modules/cerebrum/ego/llm-client.ts
@@ -7,11 +7,11 @@
  */
 import Anthropic from '@anthropic-ai/sdk';
 
-import { getEnv } from '../../env.js';
-import { withRateLimitRetry } from '../../lib/ai-retry.js';
-import { trackInference } from '../../lib/inference-middleware.js';
-import { logger } from '../../lib/logger.js';
-import { getSettingValue } from '../core/settings/service.js';
+import { getEnv } from '../../../env.js';
+import { withRateLimitRetry } from '../../../lib/ai-retry.js';
+import { trackInference } from '../../../lib/inference-middleware.js';
+import { logger } from '../../../lib/logger.js';
+import { getSettingValue } from '../../core/settings/service.js';
 
 import type { MessageStream } from '@anthropic-ai/sdk/lib/MessageStream';
