+++ apps/pops-api/src/modules/cerebrum/ego/prompts.ts
