+++ apps/pops-api/src/modules/cerebrum/ego/scope-keywords.ts
