+++ apps/pops-api/src/modules/cerebrum/ego/scope-negotiator.ts
