+++ apps/pops-api/src/modules/cerebrum/ego/settings-manifest.ts
