+++ apps/pops-api/src/modules/cerebrum/ego/types.ts
